chore(refactoring): header clean ups
Move privately used header code to cc files. Remove redunandant includes.
No functional changes.
