+++ src/server/conn_context.cc
@@ -8,6 +8,7 @@
 #include "core/heap_size.h"
 #include "facade/acl_commands_def.h"
 #include "server/acl/acl_commands_def.h"
+#include "server/channel_store.h"
 #include "server/command_registry.h"
 #include "server/engine_shard_set.h"
 #include "server/server_family.h"
